+++ Tests/ParseSwiftTests/ParseInstallationAsyncTests.swift
@@ -1119,5 +1119,75 @@ class ParseInstallationAsyncTests: XCTestCase { // swiftlint:disable:this type_b
             }
         }
     }
+
+    @MainActor
+    func testBecome() async throws {
+        try saveCurrentInstallation()
+        MockURLProtocol.removeAll()
+
+        guard let installation = Installation.current,
+            let savedObjectId = installation.objectId else {
+                XCTFail("Should unwrap")
+                return
+        }
+        XCTAssertEqual(savedObjectId, self.testInstallationObjectId)
+
+        var installationOnServer = installation
+        installationOnServer.updatedAt = installation.updatedAt?.addingTimeInterval(+300)
+        installationOnServer.customKey = "newValue"
+        installationOnServer.installationId = "wowsers"
+        installationOnServer.channels = ["yo"]
+        installationOnServer.deviceToken = "no"
+
+        let encoded: Data!
+        do {
+            encoded = try installationOnServer.getEncoder().encode(installationOnServer, skipKeys: .none)
+            // Get dates in correct format from ParseDecoding strategy
+            installationOnServer = try installationOnServer.getDecoder().decode(Installation.self, from: encoded)
+        } catch {
+            XCTFail("Should encode/decode. Error \(error)")
+            return
+        }
+        MockURLProtocol.mockRequests { _ in
+            return MockURLResponse(data: encoded, statusCode: 200, delay: 0.0)
+        }
+
+        let fetched = try await Installation.become(installationId: "wowsers")
+        guard let currentInstallation = Installation.current else {
+            XCTFail("Should have current installation")
+            return
+        }
+        XCTAssertTrue(fetched.hasSameObjectId(as: currentInstallation))
+        XCTAssertTrue(fetched.hasSameInstallationId(as: currentInstallation))
+        XCTAssertTrue(fetched.hasSameObjectId(as: installationOnServer))
+        XCTAssertTrue(fetched.hasSameInstallationId(as: installationOnServer))
+        guard let fetchedCreatedAt = fetched.createdAt else {
+            XCTFail("Should unwrap dates")
+            return
+        }
+        guard let originalCreatedAt = installationOnServer.createdAt else {
+            XCTFail("Should unwrap dates")
+            return
+        }
+        XCTAssertEqual(fetchedCreatedAt, originalCreatedAt)
+        XCTAssertEqual(fetched.channels, installationOnServer.channels)
+        XCTAssertEqual(fetched.deviceToken, installationOnServer.deviceToken)
+
+        // Should be updated in memory
+        XCTAssertEqual(Installation.current?.installationId, "wowsers")
+        XCTAssertEqual(Installation.current?.customKey, installationOnServer.customKey)
+        XCTAssertEqual(Installation.current?.channels, installationOnServer.channels)
+        XCTAssertEqual(Installation.current?.deviceToken, installationOnServer.deviceToken)
+
+        // Should be updated in Keychain
+        guard let keychainInstallation: CurrentInstallationContainer<BaseParseInstallation>
+            = try? KeychainStore.shared.get(valueFor: ParseStorage.Keys.currentInstallation) else {
+                XCTFail("Should get object from Keychain")
+            return
+        }
+        XCTAssertEqual(keychainInstallation.currentInstallation?.installationId, "wowsers")
+        XCTAssertEqual(keychainInstallation.currentInstallation?.channels, installationOnServer.channels)
+        XCTAssertEqual(keychainInstallation.currentInstallation?.deviceToken, installationOnServer.deviceToken)
+    }
 }
 #endif